Some encoder tweaks
The mouse sensor is now forced to 2000 DPI.
The incremental encoder now correctly reads a quadrature signal derived
from a common DVS timecode.

Andrea Zoppi

+++ src/firmware/include/incenc.h
@@ -47,7 +47,7 @@
 
 
 // Types
-typedef signed long				INCENC_DELTA;		///< Incremental encoder delta
+typedef signed short			INCENC_DELTA;		///< Incremental encoder delta
 
 typedef union {
 	struct {
